// interfaces/AppInterfaces.ts
export interface UsageMetadata {
    promptTokenCount:        number;
    candidatesTokenCount:    number;
    totalTokenCount:         number;
    promptTokensDetails:     TokensDetail[];
    candidatesTokensDetails: TokensDetail[];
}

export interface TokensDetail {
    modality:   string;
    tokenCount: number;
}
export interface Message {
    text: string;
    sender: string;
    timestamp: string;
    title?: string;

}